extract regexp operator to it's own file
