+++ lib/active_record/hash_options.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 require "active_record/hash_options/version"
 require "active_record/hash_options/operators"
+require "active_record/hash_options/operators/regexp"
 require "active_record/hash_options/helpers"
 require "active_record/hash_options/enumerable"
